+++ ASP.NET Core/Web/GamersHub.Web/Controllers/ReviewsController.cs
@@ -3,6 +3,7 @@ using GamersHub.Common;
 using GamersHub.Data.Models;
 using GamersHub.Services.Data.Games;
 using GamersHub.Services.Data.Reviews;
+using GamersHub.Services.Data.Users;
 using GamersHub.Web.ViewModels.Replies;
 using GamersHub.Web.ViewModels.Reviews;
 using Microsoft.AspNetCore.Authorization;
@@ -17,15 +18,18 @@ namespace GamersHub.Web.Controllers
         private readonly IGamesService gamesService;
         private readonly IReviewsService reviewsService;
         private readonly UserManager<ApplicationUser> userManager;
+        private readonly IUsersService usersService;
 
         public ReviewsController(
             IGamesService gamesService,
             IReviewsService reviewsService,
-            UserManager<ApplicationUser> userManager)
+            UserManager<ApplicationUser> userManager,
+            IUsersService usersService)
         {
             this.gamesService = gamesService;
             this.reviewsService = reviewsService;
             this.userManager = userManager;
+            this.usersService = usersService;
         }
 
         public IActionResult Create()
@@ -62,13 +66,22 @@ namespace GamersHub.Web.Controllers
             return this.RedirectToAction("ById", "Games", new {id = input.GameId, name = gameUrl});
         }
 
-        public IActionResult Edit(int id)
+        public async Task<IActionResult> Edit(int id)
         {
             var viewModel = this.reviewsService.GetById<ReviewEditViewModel>(id);
 
             if (viewModel == null)
             {
                 return this.NotFound();
+            }
+
+            var user = await this.userManager.GetUserAsync(this.User);
+
+            bool isUserAllowedToEdit = await this.usersService.ValidateUserCanEditDeleteById(viewModel.UserId, user);
+
+            if (isUserAllowedToEdit == false)
+            {
+                return this.Redirect("/Identity/Account/AccessDenied");
             }
 
             return this.View(viewModel);
@@ -78,6 +91,15 @@ namespace GamersHub.Web.Controllers
         [HttpPost]
         public async Task<IActionResult> Edit(ReviewEditViewModel input)
         {
+            var user = await this.userManager.GetUserAsync(this.User);
+
+            bool isUserAllowedToEdit = await this.usersService.ValidateUserCanEditDeleteById(input.UserId, user);
+
+            if (isUserAllowedToEdit == false)
+            {
+                return this.Redirect("/Identity/Account/AccessDenied");
+            }
+
             if (!this.ModelState.IsValid)
             {
                 return this.View(input);
@@ -93,7 +115,7 @@ namespace GamersHub.Web.Controllers
             return this.RedirectToAction("ById", "Games", new {id = input.GameId, name = input.GameUrl});
         }
 
-        public IActionResult Delete(int id)
+        public async Task<IActionResult> Delete(int id)
         {
             var viewModel = this.reviewsService.GetById<ReviewDeleteViewModel>(id);
 
@@ -102,12 +124,30 @@ namespace GamersHub.Web.Controllers
                 return this.NotFound();
             }
 
+            var user = await this.userManager.GetUserAsync(this.User);
+
+            bool isUserAllowedToDelete = await this.usersService.ValidateUserCanEditDeleteById(viewModel.UserId, user);
+
+            if (isUserAllowedToDelete == false)
+            {
+                return this.Redirect("/Identity/Account/AccessDenied");
+            }
+
             return this.View(viewModel);
         }
 
         [HttpPost]
         public async Task<IActionResult> Delete(ReviewDeleteViewModel input)
         {
+            var user = await this.userManager.GetUserAsync(this.User);
+
+            bool isUserAllowedToDelete = await this.usersService.ValidateUserCanEditDeleteById(input.UserId, user);
+
+            if (isUserAllowedToDelete == false)
+            {
+                return this.Redirect("/Identity/Account/AccessDenied");
+            }
+
             await this.reviewsService.DeleteAsync(input.Id);
 
             return this.RedirectToAction("ById", "Games", new {id = input.GameId, name = input.GameUrl});
